Stricter rating (#8)
* feat(invalid-rating): throw error when rating is out of valid range

* feat(valid-rating): rating altered in db to be integer && throw error in MADM if rating is out of range

+++ sql/create.sql
@@ -84,7 +84,7 @@ CREATE TABLE IF NOT EXISTS unit_rating (
   id SERIAL PRIMARY KEY NOT NULL,
   utari_user TEXT NOT NULL REFERENCES utari_user (id),
   unit INTEGER NOT NULL REFERENCES unit (id),
-  rating DOUBLE PRECISION NOT NULL,
+  rating INTEGER NOT NULL,
   time_created TIMESTAMPTZ NOT NULL
 );
 CREATE TABLE IF NOT EXISTS unit_bookmarked (
@@ -120,7 +120,7 @@ CREATE TABLE IF NOT EXISTS room_rating (
   id SERIAL PRIMARY KEY NOT NULL,
   utari_user TEXT NOT NULL REFERENCES utari_user (id),
   room INTEGER NOT NULL REFERENCES room (id),
-  rating DOUBLE PRECISION NOT NULL,
+  rating INTEGER NOT NULL,
   time_created TIMESTAMPTZ NOT NULL
 );
 CREATE TABLE IF NOT EXISTS room_bookmarked (
